\name{isotonic.pen-package}
\alias{isotonic.pen-package}
\alias{isotonic.pen}
\docType{package}
\title{
Penalized Isotonic Regression in one and two dimensions
}
\description{
Given a response y and a one- or two-dimensional predictor, the isotonic regression estimator is calculated with the usual orderings.  The user can specify a penalty to tame spiking, or a default value can be used. 
}
\details{
\tabular{ll}{
Package: \tab isotonic.pen\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-04\cr
License: \tab GPL-2 | GPL-3 \cr
}
}
\author{
Mary C Meyer, Jiwen Wu, and Jean D Opsomer

Maintainer: Mary C Meyer <meyer@stat.colostate.edu>
}
\references{
Meyer, M.C. (2013) A Simple New Algorithm for Quadratic Programming with Applications in Statistics, \emph{Communications in Statistics}, \bold{42(5)}, 1126-1139.
}
\keyword{ package }
\keyword{isotonic}
